+++ mtproto/src/main/kotlin/com/github/badoualy/telegram/mtproto/tl/MTPingDelayDisconnect.kt
@@ -9,7 +9,7 @@ import java.io.OutputStream
 
 class MTPingDelayDisconnect @JvmOverloads constructor(var pingId: Long = 0, var disconnectDelay: Int = 0) : TLObject() {
 
-    override fun getConstructorId(): Int {
+    override fun getConstructorId(): Long {
         return CONSTRUCTOR_ID
     }
 
@@ -31,6 +31,6 @@ class MTPingDelayDisconnect @JvmOverloads constructor(var pingId: Long = 0, var 
 
     companion object {
         @JvmField
-        val CONSTRUCTOR_ID = 213746804
+        val CONSTRUCTOR_ID = 213746804L
     }
 }
