+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TLInputMediaContact.java
@@ -66,7 +66,7 @@ public class TLInputMediaContact extends TLAbsInputMedia {
     }
 
     @Override
-    public int getConstructorId() {
+    public long getConstructorId() {
         return CONSTRUCTOR_ID;
     }
