+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TLUserEmpty.java
@@ -52,7 +52,7 @@ public class TLUserEmpty extends TLAbsUser {
     }
 
     @Override
-    public int getConstructorId() {
+    public long getConstructorId() {
         return CONSTRUCTOR_ID;
     }
